backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
httpx==0.27.0
apscheduler==3.10.4
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
python-multipart==0.0.9
